csli path bug fix
Fixed a bug that prevented clsi.exe to run from different directories. Basically the code was running the python file from the same directory of execution which is wrong we need to execute it from the path of the clsi.exe location where the python file (python interpreter) is found.